requirements.txt
Flask
azure-storage-blob
requests
python-dotenv
